+++ src/QuranImageMaker/OutputStorageLocation.cs
@@ -1,9 +1,23 @@
 ﻿namespace QuranImageMaker
 {
+    /// <summary>
+    /// Specifies the type of output for the Quran image maker.
+    /// </summary>
     public enum OutputType
     {
+        /// <summary>
+        /// Output as a bitmap image.
+        /// </summary>
         Bitmap,
+
+        /// <summary>
+        /// Output as a byte array.
+        /// </summary>
         Bytes,
+
+        /// <summary>
+        /// Output as a file.
+        /// </summary>
         File
     }
 }
