+++ src/QuranImageMaker/TextPaints.cs
@@ -1,0 +1,20 @@
+﻿using SkiaSharp;
+
+namespace QuranImageMaker
+{
+    /// <summary>
+    /// Represents the text paints used for rendering.
+    /// </summary>
+    public class TextPaints
+    {
+        /// <summary>
+        /// Gets or sets the text paint.
+        /// </summary>
+        public SKPaint TextPaint { get; set; }
+
+        /// <summary>
+        /// Gets or sets the text shadow paint.
+        /// </summary>
+        public SKPaint TextShadowPaint { get; set; }
+    }
+}
