+++ src/QuranImageMaker/TextPosition.cs
@@ -1,9 +1,23 @@
 ﻿namespace QuranImageMaker
 {
+    /// <summary>
+    /// Specifies the position of the text within an image.
+    /// </summary>
     public enum TextPosition
     {
+        /// <summary>
+        /// The text is centered within the image.
+        /// </summary>
         Center,
+
+        /// <summary>
+        /// The text is positioned at the top of the image.
+        /// </summary>
         Top,
+
+        /// <summary>
+        /// The text is positioned at the bottom of the image.
+        /// </summary>
         Bottom,
     }
 }
